Better alternative auth scheme support in resource server.
Improved support for alternative authentication schemes on
the resource server side, most notably for the MAC scheme,
which also requires different use of attributes in the
WWW-Authenticate header.

By only having the MacAuth in the list of supported schemes,
it's now possible to inform clients correctly about the
required authentication scheme, instead of always returning
the "Bearer" challenge when no token was provided at all.

Also, adding some additional tests to ensure we are in line
with the current specs for OAuth2 as well as Bearer and Mac
schemes.