Replace universal servlet response wrapper with an explict option.
Universally wrapping the HttpServletResponse in order to support its
native Writer interface conflicts with the generic ResponseFilter
paradigm (mostly applied in unfiltered.kit). A ResponseFilter must be
able to replace the outputStream that is made available to the
response functions that follow. This replacement reference is not
available to a ResponseWrapper that was instantiated before any of the
response functions were applied. And so, the filtered stream is not
the one that was closed by the plan and filtered responses were
incorrect in some cases.

This second approach requires a client application to explicitly wrap
the underlying response if it intends to use it with software that
uses its Writer interface. It is also assumed that the client
application will close the Writer when it is finished.