Continue to use GlobalEventExecutor as default in netty-server.
I brought this up as a comment in the dropped pr #240 but forgot to
circle back to it in #242.

I wasn't convinced that `GlobalEventExecutor` is a bad default, and
I'm pretty sure that `ImmediateEventExecutor` "which executes tasks in
the callers thread" is a worse one. As mentioned in this
[stackoverflow thread][so] which suggests that GEE is a reasonable
choice, this executor is used "to notify the ChannelGroupFutures". I
don't think it's a good idea to block the calling thread to do that.

We aren't limited to the 3 executors included with Netty but among
those `GlobalEventExecutor` seems to be the most appropriate. For now
let's stick with that. With the refactoring, anybody can supply their
own executor and I'm making that a little easier here
too (`channelsExecutor` builder method).

[so]: http://stackoverflow.com/a/18072390